requirements.txt
numpy==1.23.5
open3d==0.19.0
Pillow==11.2.1
scikit_learn==1.7.0
scipy==1.16.0
torch==2.7.1
torch_cluster==1.6.3+pt27cu126
tqdm==4.67.1
